Fix offset arg type being int, now it's float

+++ main_scripts/map_buildings_to_tiles.py
@@ -64,8 +64,8 @@ def main():
     parser.add_argument("-c", "--city-name", required=True, help="Name of city (sub-directory of output directory that will be created)")
     parser.add_argument("--sw", required=True, help='SW corner formatted as "lat,lon" or "lat, lon"')
     parser.add_argument("--ne", required=True, help='NE corner formatted as "lat,lon" or "lat, lon"')
-    parser.add_argument("--offset-x", required=False, type=int, default=0, help='Offset x coord of each point')
-    parser.add_argument("--offset-y", required=False, type=int, default=0, help='Offset y coord of each point')
+    parser.add_argument("--offset-x", required=False, type=float, default=0., help='Offset x coord of each point')
+    parser.add_argument("--offset-y", required=False, type=float, default=0., help='Offset y coord of each point')
 
     args = parser.parse_args()
